Semi-Random zones
ability to specify which face(s) the zone should be able to spawn on. Useful for showing the zone on your BlockFace when initially hit.

Need to optimize GetTree() making a new Tree() to compare against when every new Tree() makes a TreeZone() and that does a bunch of calculations that is unneeded for just a simple comparison. Make it so TreeZone needs to be generated by a void.